src/requirements.txt
attrs==17.4.0
certifi==2018.1.18
chardet==3.0.4
coverage==4.5.1
decorator==4.2.1
Django==2.0.2
django-environ==0.4.4
django-extensions==2.0.0
djangorestframework==3.7.7
flake8==3.5.0
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.11.1
mccabe==0.6.1
parso==0.1.1
pexpect==4.4.0
django-filter==2.0.0
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
psycopg2==2.7.5
ptyprocess==0.5.2
py==1.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pytest==3.4.1
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-flake8==0.9.1
pytest-mock==1.7.1
pytz==2018.3
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
typing==3.6.4
urllib3==1.22
wcwidth==0.1.7
